Fifa Player Wage Prediction
This is an analysis of the game FIFA 18, focusing on predicting the
players wage based on the features available in the FIFA 2018 dataset.
This prediction problem falls under the sports domain. I am aiming to
fit a model that can predict all players wages appropriately. The
CompleteDataset is downloaded from Kaggle.